Raise exceptions when no handlers are provided
Previously, writing a bare (try (foo)) would invoke Pokemon exception catching (gotta catch 'em all) instead of the correct behavior, which is to raise the exception if no handler is provided.

Note that this is a cute feature of Hy, as a `try` with no `except` is a syntax error.  We avoid the syntax error here because we don't use Python's compiler, which is the only thing that can throw Syntax Errors.  :D