Avoid compliation of AST builtins during a Call
Since it defers back to compile expression, it'll defer to @builds("+"), which
means we get a BinOp back. Which is *hilarious*.  At the top there should be a
self.compile call since it'll actually treat the first argument as the argument
in the call position when it needs to be an ast.Call.