Replace `true` with `T` and `false` with `F`
Per discussion in #908.

True and False are still allowed. Allowing users to use these as variable names is probably more trouble than it's worth, and I figure this helps to smooth over human translation from Python code to Hy. It might be a good idea to use T and F consistently in Hy's own code, not to mention the documentation. (This preference could be stated in the style guide.)

I added a bit to the documentation about the correspondences between True, False, None, and their synonyms because this didn't seem to be documented at all.