Replace the loader with a more compatable implementation
Add the necessary bits to provide a Python 3.4+ compliant import
machinery. This means we now have our own PathFinder, which we register
as in sys.meta_path, our own FileFinder, and our own Loader.

Add a compatability later that builds on the Python 3 machinery to make
it best-effort backwards compatible with Python 2.