Improve `replace` on HyObject so it does less work
Calling `replace` can trigger a call to `replace_hy_obj`, which will
try to also wrap values inside a `HyObject`.

However, this means when we try to recusively replace inside a nested
object, like a `HyList`, we should try and capture these new wrapped
objects.

Hy still works if we don't, but it results in the object getting
rewrapped multiple times during a compile.

Closes #1519