In `test-doc`, use a purpose-written macro
This prevents an issue where the test fails when it's run from bytecode because `__macros__` doesn't exist. It also lets us test that `doc` gets the docstring without making the test dependent on the docstring of `doc` itself.

I haven't tried to fix `echo '(doc doc)' | hy2py | python3`. That will still fail. But I think that's okay because `doc` is intended for interactive use, not for programs.