Added support for match statement from python 3.10
Patterns are a dsl within Python. Hence this affects the compilations
of other forms within a pattern. `(:as ...)` and `(:or ...)` are used
as the as and or patterns. Special forms are ignored.

If the guard of a match case has statements, the match form is split
into nested match forms to execute them.