Use "environment markers" for conditional install_requires
These are defined in PEP 508: https://www.python.org/dev/peps/pep-0508/#environment-markers

The current implementation means that installing 1.0a2 from the published wheel on Python 3.8 does not include `astor` (and thus does not work) because "the Python version that is building the wheel is different from the Python version that is installing it" (from https://hynek.me/articles/conditional-python-dependencies/).

I don't think Windows and Linux can share wheels, so `pyreadline` doesn't strictly have to be included here, but it seems more correct to use this method for both.