Remove the docstring of `hy.compiler.hy_eval`
It's not quite correct. It may've been once, but it hasn't been for a while (e.g., `read-str` is long gone). Time that might be spent on making sure the description is correct would be better spent on cleaning up the implementation. The larger issue is that we already have a lot of undermaintained user documentation; purely internal documentation is an additional maintenance burden.