Overhaul debugging environment variables
`HY_DEBUG` and `HY_FILTER_INTERNAL_ERRORS` were redundant in intention, but `HY_DEBUG` worked more reliably because `hy_exc_handler` is called explicitly in a few places. The new environment variable `HY_SHOW_INTERNAL_ERRORS` uses a more specific name while avoiding the awkwardness of how `HY_FILTER_INTERNAL_ERRORS` had to be set to an empty string to have an effect.

The new environment variable is still untested, because it's only for internal use, and is documented merely for completeness.